pyproject.toml
[project]
description = "Reinforcement Learning driven simulation of vacancy diffusion"
name = "rlsim"
readme = "README.md"

classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
  "numpy>=1.22.4",
  "scipy>=1.7.1",
  "matplotlib>=3.4.3",
  "ase>=3.22.1",
  "torch>=1.11.0",
  "networkx>=2.7.1",
  "pymatgen>=2022.5.26",
  "torch>=1.12.0",
  "torchdata>=0.6.0",
  "torch-geometric>=2.3",
  "torch_sparse",
  "inflection",
  "torch_scatter",
  "torch_cluster",
  "e3nn==0.4.4",
  "torch-ema",
  "rich",
  "tensorboard",
  "matscipy",
  "ase-extension",
  "jupyter",
  "mace-torch",
]
dynamic = ["version"]
license = {text = "MIT"}
requires-python = ">=3.9,<3.12"

# [project.scripts]
# rlsim = "vssr.cli.main:main"

[project.optional-dependencies]
dev = ["black", "isort", "flake8"]

[tool.setuptools.packages.find]
exclude = ["dev"]
include = ["rlsim"]

[tool.black]
line-length = 120

[tool.setuptools.dynamic]
version = {attr = "rlsim.__version__"}

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "setuptools-scm"]
